+++ hugr-py/src/hugr/cfg.py
@@ -1,10 +1,14 @@
+"""Builder classes for HUGR control flow graphs."""
+
 from __future__ import annotations
 
+from contextlib import AbstractContextManager
 from dataclasses import dataclass
 from typing import TYPE_CHECKING
 
-import hugr.ops as ops
-import hugr.val as val
+from typing_extensions import Self
+
+from hugr import ops, val
 
 from .dfg import _DfBase
 from .exceptions import MismatchedExit, NoSiblingAncestor, NotInSameCfg
@@ -16,6 +20,8 @@ if TYPE_CHECKING:
 
 
 class Block(_DfBase[ops.DataflowBlock]):
+    """Builder class for a basic block in a HUGR control flow graph."""
+
     def set_block_outputs(self, branching: Wire, *other_outputs: Wire) -> None:
         self.set_outputs(branching, *other_outputs)
 
@@ -44,16 +50,33 @@ class Block(_DfBase[ops.DataflowBlock]):
 
 
 @dataclass
-class Cfg(ParentBuilder[ops.CFG]):
+class Cfg(ParentBuilder[ops.CFG], AbstractContextManager):
+    """Builder class for a HUGR control flow graph, with the HUGR root node
+    being a :class:`CFG <hugr.ops.CFG>`.
+
+    Args:
+        input_types: The input types for the CFG. Outputs are computed
+        by propagating types through the control flow graph to the exit block.
+
+    Examples:
+        >>> cfg = Cfg(tys.Bool, tys.Unit)
+        >>> cfg.parent_op
+        CFG(inputs=[Bool, Unit])
+    """
+
+    #: The HUGR instance this CFG is part of.
     hugr: Hugr
+    #: The parent node of the CFG.
     parent_node: Node
     _entry_block: Block
+    #: The node holding the root of the exit block.
     exit: Node
 
-    def __init__(self, input_types: TypeRow) -> None:
-        root_op = ops.CFG(inputs=input_types)
+    def __init__(self, *input_types: Type) -> None:
+        input_typs = list(input_types)
+        root_op = ops.CFG(inputs=input_typs)
         hugr = Hugr(root_op)
-        self._init_impl(hugr, hugr.root, input_types)
+        self._init_impl(hugr, hugr.root, input_typs)
 
     def _init_impl(self: Cfg, hugr: Hugr, root: Node, input_types: TypeRow) -> None:
         self.hugr = hugr
@@ -63,6 +86,12 @@ class Cfg(ParentBuilder[ops.CFG]):
 
         self.exit = self.hugr.add_node(ops.ExitBlock(), self.parent_node)
 
+    def __enter__(self) -> Self:
+        return self
+
+    def __exit__(self, *args) -> None:
+        return None
+
     @classmethod
     def new_nested(
         cls,
@@ -70,6 +99,23 @@ class Cfg(ParentBuilder[ops.CFG]):
         hugr: Hugr,
         parent: ToNode | None = None,
     ) -> Cfg:
+        """Start building a CFG nested inside an existing HUGR graph.
+
+        Args:
+            input_types: The input types for the CFG.
+            hugr: The HUGR instance this CFG is part of.
+            parent: The parent node for the CFG: defaults to the root of the HUGR
+                instance.
+
+        Returns:
+            The new CFG builder.
+
+        Examples:
+            >>> hugr = Hugr()
+            >>> cfg = Cfg.new_nested([tys.Bool], hugr)
+            >>> cfg.parent_op
+            CFG(inputs=[Bool])
+        """
         new = cls.__new__(cls)
         root = hugr.add_node(
             ops.CFG(inputs=input_types),
@@ -80,6 +126,13 @@ class Cfg(ParentBuilder[ops.CFG]):
 
     @property
     def entry(self) -> Node:
+        """Node for entry block of the CFG.
+
+        Examples:
+            >>> cfg = Cfg(tys.Bool)
+            >>> cfg.entry
+            Node(1)
+        """
         return self._entry_block.parent_node
 
     @property
@@ -91,11 +144,34 @@ class Cfg(ParentBuilder[ops.CFG]):
         return self.hugr._get_typed_op(self.exit, ops.ExitBlock)
 
     def add_entry(self) -> Block:
+        """Start building the entry block of the CFG.
+
+        Returns:
+            The entry block builder.
+
+        Examples:
+            >>> cfg = Cfg(tys.Bool)
+            >>> entry = cfg.add_entry()
+            >>> entry.set_outputs(*entry.inputs())
+        """
         return self._entry_block
 
-    def add_block(self, input_types: TypeRow) -> Block:
+    def add_block(self, *input_types: Type) -> Block:
+        """Add a new block to the CFG and start building it.
+
+        Args:
+            input_types: The input types for the block.
+
+        Returns:
+            The block builder.
+
+        Examples:
+            >>> cfg = Cfg(tys.Bool)
+            >>> with cfg.add_block(tys.Unit) as b:\
+                    b.set_single_succ_outputs(*b.inputs())
+        """
         new_block = Block.new_nested(
-            ops.DataflowBlock(input_types),
+            ops.DataflowBlock(list(input_types)),
             self.hugr,
             self.parent_node,
         )
@@ -104,7 +180,24 @@ class Cfg(ParentBuilder[ops.CFG]):
     # TODO insert_block
 
     def add_successor(self, pred: Wire) -> Block:
-        b = self.add_block(self._nth_outputs(pred))
+        """Start building a block that succeeds an existing block.
+
+        Args:
+            pred: The wire from the predecessor block to the new block. The
+            port of the wire determines the branching index of the new block.
+
+
+        Returns:
+            The new block builder.
+
+        Examples:
+            >>> cfg = Cfg(tys.Bool)
+            >>> with cfg.add_entry() as entry:\
+                    entry.set_single_succ_outputs()
+            >>> with cfg.add_successor(entry[0]) as b:\
+                    b.set_single_succ_outputs(*b.inputs())
+        """
+        b = self.add_block(*self._nth_outputs(pred))
 
         self.branch(pred, b)
         return b
@@ -115,6 +208,19 @@ class Cfg(ParentBuilder[ops.CFG]):
         return block.nth_outputs(port.offset)
 
     def branch(self, src: Wire, dst: ToNode) -> None:
+        """Add a branching control flow link between blocks.
+
+        Args:
+            src: The wire from the predecessor block.
+            dst: The destination block.
+
+        Examples:
+            >>> cfg = Cfg(tys.Bool)
+            >>> with cfg.add_entry() as entry:\
+                    entry.set_single_succ_outputs()
+            >>> b = cfg.add_block(tys.Unit)
+            >>> cfg.branch(entry[0], b)
+        """
         # TODO check for existing link/type compatibility
         if dst.to_node() == self.exit:
             return self.branch_exit(src)
@@ -122,6 +228,17 @@ class Cfg(ParentBuilder[ops.CFG]):
         self.hugr.add_link(src, dst.inp(0))
 
     def branch_exit(self, src: Wire) -> None:
+        """Branch from a block to the exit block.
+
+        Args:
+            src: The wire from the predecessor block.
+
+        Examples:
+            >>> cfg = Cfg(tys.Bool)
+            >>> with cfg.add_entry() as entry:\
+                    entry.set_single_succ_outputs()
+            >>> cfg.branch_exit(entry[0])
+        """
         src = src.out_port()
         self.hugr.add_link(src, self.exit.inp(0))