+++ hugr-py/src/hugr/serialization/__init__.py
@@ -1,0 +1,1 @@
+"""Serialized HUGR objects as pydantic models."""